moved chat config into chat category and changed to red color for incompatibility message in config description.
