+++ src/main/java/at/hannibal2/skyhanni/features/misc/PlayerChatSymbols.kt
@@ -16,7 +16,7 @@ import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 
 // code inspired by SBA but heavily modified to be more functional and actually work
 class PlayerChatSymbols {
-    private val config get() = SkyHanniMod.feature.misc.chatSymbols
+    private val config get() = SkyHanniMod.feature.chat.chatSymbols
     private val nameSymbols = mutableMapOf<String, String>()
     private val symbolsPattern = "^(?<symbols>(?:(?:§\\w)+\\S)+) ".toPattern()
     private val symbolPattern = "((?:§\\w)+\\S)".toPattern()
